@@ -9,12 +9,12 @@
   "dependencies": {
     "axios": "^0.21.1",
     "promise-polyfill": "^8.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2",
+    "react-dom": "^18.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "ts-loader": "^9.1.1",
     "typescript": "^4.2.4",
     "webpack": "^5.35.1",
